Move the static files from the original expense tracker into app template
