+++ expensetracker/core/static/app/style.css
