--- index.html
@@ -1,88 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Expense Tracker</title>
-    <link rel="stylesheet" href="style.css" />
-    <link rel="stylesheet" href="reset.css" />
-  </head>
-  <body>
-    <div class="sidebar">
-      <h2>Add new transaction</h2>
-      <form id="add-transaction-form" class="add-transaction-form">
-        <div class="input-group">
-          <label for="transact-text">Text:</label>
-          <input
-            type="text"
-            id="transact-text"
-            name="transact-text"
-            placeholder="Enter text..."
-          />
-          <object
-            id="success-svg"
-            class="icon icon-success hidden"
-            data="./image/success.svg"
-            type="image/svg+xml"
-          ></object>
-          <object
-            id="error-svg"
-            class="icon icon-error hidden"
-            data="./image/error.svg"
-            type="image/svg+xml"
-          ></object>
-          <span class="error-message"></span>
-        </div>
-        <div class="input-group">
-          <label for="transact-amount">Amount:</label>
-          <input
-            type="text"
-            id="transact-amount"
-            name="transact-amount"
-            placeholder="Enter amount..."
-          />
-          <object
-            id="success-svg"
-            class="icon icon-success hidden"
-            data="./image/success.svg"
-            type="image/svg+xml"
-          ></object>
-          <object
-            id="error-svg"
-            class="icon icon-error hidden"
-            data="./image/error.svg"
-            type="image/svg+xml"
-          ></object>
-          <span class="error-message"></span>
-        </div>
-        <label>Category:</label>
-        <div class="radio" id="radio"></div>
-        <input type="submit" value="Submit" id="submit" disabled />
-      </form>
-    </div>
-    <main>
-      <section class="balance-section">
-        <h2>YOUR BALANCE</h2>
-        <p id="balance" class="balance"></p>
-        <div class="account-summary">
-          <div class="income" id="income">
-            <h3>INCOME</h3>
-            <p></p>
-          </div>
-          <div class="expenses" id="expenses">
-            <h3>EXPENSES</h3>
-            <p></p>
-          </div>
-        </div>
-      </section>
-      <section class="history-section">
-        <h2>HISTORY</h2>
-        <div class="history-card-container" id="history-card-container"></div>
-      </section>
-    </main>
-
-    <script type="module" src="script.js"></script>
-    <script type="module" src="validation.js"></script>
-  </body>
-</html>
