+++ templates/core/app.html
@@ -1,7 +1,87 @@
-{% if user.is_authenticated %} Hi {{user.email}}!
-<p><a href="{%url 'account_logout' %}">Log Out </a></p>
-{%else%}
+{% load static %}
+
+<link rel="stylesheet" type="text/css" href="{% static 'app/style.css' %}" />
+<link rel="stylesheet" type="text/css" href="{% static 'app/reset.css' %}" />
+
+{% if user.is_authenticated %} Hi {{ user.email }}!
+<body>
+  <div class="sidebar">
+    <p><a href="{% url 'account_logout' %}">Log Out</a></p>
+    <h2>Add new transaction</h2>
+    <form id="add-transaction-form" class="add-transaction-form">
+      <div class="input-group">
+        <label for="transact-text">Text:</label>
+        <input
+          type="text"
+          id="transact-text"
+          name="transact-text"
+          placeholder="Enter text..."
+        />
+        <object
+          id="success-svg"
+          class="icon icon-success hidden"
+          data="{% static 'app/images/success.svg'%}"
+          type="image/svg+xml"
+        ></object>
+        <object
+          id="error-svg"
+          class="icon icon-error hidden"
+          data="{% static 'app/images/error.svg'%}"
+          type="image/svg+xml"
+        ></object>
+        <span class="error-message"></span>
+      </div>
+      <div class="input-group">
+        <label for="transact-amount">Amount:</label>
+        <input
+          type="text"
+          id="transact-amount"
+          name="transact-amount"
+          placeholder="Enter amount..."
+        />
+        <object
+          id="success-svg"
+          class="icon icon-success hidden"
+          data="{% static 'app/images/success.svg'%}"
+          type="image/svg+xml"
+        ></object>
+        <object
+          id="error-svg"
+          class="icon icon-error hidden"
+          data="{% static 'app/images/error.svg'%}"
+          type="image/svg+xml"
+        ></object>
+        <span class="error-message"></span>
+      </div>
+      <label>Category:</label>
+      <div class="radio" id="radio"></div>
+      <input type="submit" value="Submit" id="submit" disabled />
+    </form>
+  </div>
+  <main>
+    <section class="balance-section">
+      <h2>YOUR BALANCE</h2>
+      <p id="balance" class="balance"></p>
+      <div class="account-summary">
+        <div class="income" id="income">
+          <h3>INCOME</h3>
+          <p></p>
+        </div>
+        <div class="expenses" id="expenses">
+          <h3>EXPENSES</h3>
+          <p></p>
+        </div>
+      </div>
+    </section>
+    <section class="history-section">
+      <h2>HISTORY</h2>
+      <div class="history-card-container" id="history-card-container"></div>
+    </section>
+  </main>
+  <script src="{% static 'app/script.js' %}"></script>
+</body>
+{% else %}
 <p>You are not logged in</p>
-<a href="{% url 'account_login'%}">Log In</a>
-<a href="{% url 'account_signup'%}">Sign Up</a>
+<a href="{% url 'account_login' %}">Log In</a> |
+<a href="{% url 'account_signup' %}">Sign Up</a>
 {% endif %}
