+++ templates/core/index.html
@@ -1,8 +1,8 @@
-<h1>Expense Tracker App</h1>
-{% if user.is_authenticated %} Hi {{ user.email }}!
-<p><a href="{% url 'account_logout' %}">Log Out</a></p>
-{% else %}
+<h1>Expense Tracker</h1>
+{% if user.is_authenticated %} Hi {{user.email}}!
+<p><a href="{%url 'account_logout' %}">Log Out </a></p>
+{%else%}
 <p>You are not logged in</p>
-<a href="{% url 'account_login' %}">Log In</a> |
-<a href="{% url 'account_signup' %}">Sign Up</a>
+<a href="{% url 'account_login'%}">Log In</a>
+<a href="{% url 'account_signup'%}">Sign Up</a>
 {% endif %}
